Added additional buttons (preliminary version)
The keys for them do not work, and they just use the settings of the
existing buttons. This is just a proof of concept.  Did have to pass
a partial of the Reviewer._answerCard() to scheduler.answer() to get
them to work. _answerCard() does not like any ease above 4.  Used a
partial because the scheduler is a different class.